requirements.txt
# annotated-types==0.7.0
# anyio==4.9.0
# click==8.1.8
# coloredlogs==15.0.1
# exceptiongroup==1.2.2
# fastapi==0.115.12
# flatbuffers==25.2.10
# h11==0.14.0
# humanfriendly==10.0
# idna==3.10
# mpmath==1.3.0
# numpy==2.0.2
# onnxruntime==1.19.2
# opencv-python-headless==4.11.0.86
# packaging==24.2
# protobuf==6.30.2
# pydantic==2.11.3
# pydantic_core==2.33.1
# python-multipart==0.0.20
# sniffio==1.3.1
# starlette==0.46.2
# sympy==1.13.3
# typing-inspection==0.4.0
# typing_extensions==4.13.2
# uvicorn==0.34.1
# pillow==11.2.1

fastapi==0.110.0
uvicorn[standard]==0.29.0
onnxruntime==1.17.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
pillow==10.3.0
psutil==5.9.8
python-multipart==0.0.9